EMR/requirements.txt
asgiref==3.8.1
babel==2.17.0
Django==5.2.1
django-cors-headers==4.7.0
django-money==3.5.4
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
packaging==25.0
pillow==11.3.0
py-moneyed==3.0
PyJWT==2.9.0
python-dotenv==1.1.0
setuptools==80.9.0
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
whitenoise==6.9.0
